requirements.txt
streamlit
geopandas
pandas
folium
streamlit-folium
shapely
pyproj
fiona
python-docx
lxml
